+++ public/app/features/trails/Integrations/DataTrailEmbedded.tsx
