+++ public/app/features/trails/Integrations/dashboardIntegration.ts
