+++ public/app/features/trails/Integrations/getQueryMetrics.ts
