+++ public/app/features/trails/Integrations/logsIntegration.ts
