+++ public/app/features/trails/Integrations/testData/no-rules.json
