+++ public/app/features/trails/Integrations/testData/with-rules.json
